# backend/apidoc/openapi.yaml
openapi: 3.0.0
info:
  description: API Documentation for Meal Planner App
  version: 1.0.0
  title: Meal Planner APIDoc
servers:
  - url: api
...
paths:
...
  /ingredients:
    post:
      tags:
        - ingredients
      summary: Add Ingredient
      description: Adds an Ingredient
      operationId: addIngredient
      requestBody:
        description: An Ingredient Object
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Rosie"
                macros:
                  type: object
                  properties:
                    protein:
                      type: number
                      format: float
                      example: 30.2
                    carbs:
                      type: number
                      format: float
                      example: 50.5
                    fat:
                      type: number
                      format: float
                      example: 20.5
                    fiber:
                      type: number
                      format: float
                      example: 5.1
                    sugar:
                      type: number
                      format: float
                      example: 15.3
                    saturated:
                      type: number
                      format: float
                      example: 5.3
                    polyunsaturated:
                      type: number
                      format: float
                      example: 3.2
                    monounsaturated:
                      type: number
                      format: float
                      example: 2.0
                    trans:
                      type: number
                      format: float
                      example: 0
                  required:
                    - protein
                    - carbs
                    - fat
                micros:
                  type: object
                  properties:
                    sodium:
                      type: number
                      format: float
                      example: 2300
                    potassium:
                      type: number
                      format: float
                      example: null
                    vitaminA:
                      type: number
                      format: float
                      example: null
                    vitaminC:
                      type: number
                      format: float
                      example: null
                    calcium:
                      type: number
                      format: float
                      example: null
                    iron:
                      type: number
                      format: float
                      example: null
              required:
                - name
                - macros
      responses:
        "200":
          description: OK. Ingredient succesfully added to database
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "http://localhost:8080/api/ingredients/3PXW8J1QGkQlZgnjOaKv"
                  message:
                    type: string
                    example: "Ingredient successfully added to database"
        "400":
          description: Bad Request. Incomplete Request Body
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Body should contain name and macros!"
        "409":
          description: Conflict. Ingredient already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ingredient already exists"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
...
  /ingredients/{ingredientUID}/unit/{unit}:
    get:
      tags:
        - ingredients
      summary: Get Ingredient Unit of Ingredient
      description: Get Ingredient Unit of Ingredient
      operationId: getIngredientUnitOfIngredient
      parameters:
        - in: path
          name: ingredientUID
          schema:
            type: string
          required: true
          description: The UID of the Ingredient
        - in: path
          name: unit
          schema:
            type: string
          required: true
          description: The Name of the Unit
      responses:
        "200":
          description: OK. Ingredient Unit retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      miligrams:
                        type: integer
                        example: 100
                  message:
                    type: string
                    example: "Ingredient Unit successfully retrieved"
        "404":
          description: Not Found. Ingredient Unit does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ingredient Unit does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
    put:
      tags:
        - ingredients
      summary: Update Ingredient Unit By UID
      description: Updates Ingredient Unit by UID
      operationId: updateIngredientUnitByUID
      parameters:
        - in: path
          name: ingredientUID
          schema:
            type: string
          required: true
          description: The UID of the Ingredient
        - in: path
          name: unit
          schema:
            type: string
          required: true
          description: The Name of the Unit
      responses:
        "200":
          description: OK. Ingredient Unit updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "http://localhost:8080/api/ingredients/3PXW8J1QGkl5gnjOaKvw/unit/cup"
                  message:
                    type: string
                    example: "Ingredient Unit successfully updated"
        "404":
          description: Not Found. Ingredient Unit does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ingredient Unit does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
    delete:
      tags:
        - ingredients
      summary: Delete Ingredient Unit By UID
      description: Delete Ingredient Unit By UID
      operationId: deleteIngredientUnitByUID
      parameters:
        - in: path
          name: ingredientUID
          schema:
            type: string
          required: true
          description: The UID of the Ingredient
        - in: path
          name: unit
          schema:
            type: string
          required: true
          description: The Name of the Unit
      responses:
        "200":
          description: OK. Ingredient Unit deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ingredient Unit successfully deleted"
        "404":
          description: Not Found. Ingredient Unit does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ingredient Unit does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /ingredients/{ingredientName}:
    get:
      tags:
        - ingredients
      summary: Get Ingredient By Name
      description: Gets Ingredient UID by Name
      operationId: getIngredientByName
      parameters:
        - in: path
          name: ingredientName
          schema:
            type: string
          required: true
          description: The Name of the Ingredient
      responses:
        "200":
          description: OK. Ingredient succesfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "http://localhost:8080/api/ingredients/DoARzpyNG8mx5Ke6bvnM"
                  message:
                    type: string
                    example: "Ingredient successfully retrieved"
        "404":
          description: Not Found. Ingredient does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ingredient does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /ingredients/{ingredientUID}:
    get:
      tags:
        - ingredients
      summary: Get Ingredient By UID
      description: Gets Ingredient Info by UID
      operationId: getIngredientByUID
      parameters:
        - in: path
          name: ingredientUID
          schema:
            type: string
          required: true
          description: The UID of the Ingredient
      responses:
        "200":
          description: OK. Ingredient succesfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      name:
                        type: string
                        example: "Malai"
                      macros:
                        type: object
                        properties:
                          protein:
                            type: number
                            format: float
                            example: 30.2
                          carbs:
                            type: number
                            format: float
                            example: 50.5
                          fat:
                            type: number
                            format: float
                            example: 20.5
                          fiber:
                            type: number
                            format: float
                            example: 5.1
                          sugar:
                            type: number
                            format: float
                            example: 15.3
                          saturated:
                            type: number
                            format: float
                            example: 5.3
                          polyunsaturated:
                            type: number
                            format: float
                            example: 3.2
                          monounsaturated:
                            type: number
                            format: float
                            example: 2.0
                          trans:
                            type: number
                            format: float
                            example: 0
                      micros:
                        type: object
                        properties:
                          sodium:
                            type: number
                            format: float
                            example: 2300
                          potassium:
                            type: number
                            format: float
                            example: null
                          vitaminA:
                            type: number
                            format: float
                            example: null
                          vitaminC:
                            type: number
                            format: float
                            example: null
                          calcium:
                            type: number
                            format: float
                            example: null
                          iron:
                            type: number
                            format: float
                            example: null
                  message:
                    type: string
                    example: "Ingredient successfully retrieved"
        "404":
          description: Not Found. Ingredient does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
    put:
      tags:
        - ingredients
      summary: Update Ingredient By UID
      description: Update Ingredient Info By UID
      operationId: updateIngredientByUID
      parameters:
        - in: path
          name: ingredientUID
          schema:
            type: string
          required: true
          description: The UID of the Ingredient
      requestBody:
        description: Ingredient Object
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Rosie"
                macros:
                  type: object
                  properties:
                    protein:
                      type: number
                      format: float
                      example: 30.2
                    carbs:
                      type: number
                      format: float
                      example: 50.5
                    fat:
                      type: number
                      format: float
                      example: 20.5
                    fiber:
                      type: number
                      format: float
                      example: 5.1
                    sugar:
                      type: number
                      format: float
                      example: 15.3
                    saturated:
                      type: number
                      format: float
                      example: 5.3
                    polyunsaturated:
                      type: number
                      format: float
                      example: 3.2
                    monounsaturated:
                      type: number
                      format: float
                      example: 2.0
                    trans:
                      type: number
                      format: float
                      example: 0
                  required:
                    - protein
                    - carbs
                    - fat
                micros:
                  type: object
                  properties:
                    sodium:
                      type: number
                      format: float
                      example: 2300
                    potassium:
                      type: number
                      format: float
                      example: null
                    vitaminA:
                      type: number
                      format: float
                      example: null
                    vitaminC:
                      type: number
                      format: float
                      example: null
                    calcium:
                      type: number
                      format: float
                      example: null
                    iron:
                      type: number
                      format: float
                      example: null
      responses:
        "200":
          description: OK. Ingredient succesfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "http://localhost:8080/api/ingredients/3PXW8J1QGkQlZgnjOaKv"
                  message:
                    type: string
                    example: "Ingredient successfully updated!"
        "404":
          description: Not Found. Ingredient does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ingredient does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
    delete:
      tags:
        - ingredients
      summary: Delete Ingredient By UID
      description: Delete Ingredient By UID
      operationId: deleteIngredientByUID
      parameters:
        - in: path
          name: ingredientUID
          schema:
            type: string
          required: true
          description: The UID of the Ingredient
      responses:
        "200":
          description: OK. Ingredient succesfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ingredient successfully deleted!"
        "404":
          description: Not Found. Ingredient does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ingredient does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
...
  /tags/{tagUID}:
    get:
      tags:
        - tags
      summary: Get Tag By UID
      description: Gets a Tag By UID
      operationId: getTagByUID
      parameters:
        - in: path
          name: tagUID
          schema:
            type: string
          required: true
          description: The UID of the Tag
      responses:
        "200":
          description: OK. Tag succesfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      text:
                        type: string
                        example: "Text"
                  message:
                    type: string
                    example: "Tag successfully retrieved"
        "400":
          description: Bad Request. Incomplete Request Body
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Body should contain text!"
        "409":
          description: Conflict. Tag already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Tag already exists"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
    delete:
      tags:
        - tags
      summary: Delete Tag By UID
      description: Delete Tag By UID
      operationId: deleteIngredientByUID
      parameters:
        - in: path
          name: tagUID
          schema:
            type: string
          required: true
          description: The UID of the Tag
      responses:
        "200":
          description: OK. Tag succesfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Tag successfully deleted!"
        "404":
          description: Not Found. Tag does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Tag does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
...